+++ web-test-runner.config.js
@@ -9,7 +9,19 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
+
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { defaultReporter } from '@web/test-runner';
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { junitReporter } from '@web/test-runner-junit-reporter';
+
 export default {
+  nodeResolve: true,
+  coverage: true,
+  reporters: [
+    defaultReporter(),
+    junitReporter(),
+  ],
   testFramework: {
     type: 'mocha',
     config: {
@@ -23,6 +35,7 @@ export default {
       'test/fixtures/**',
       'node_modules/**',
       '.rum/**',
+      'src/index.js',
     ],
   },
   files: [
@@ -34,7 +47,25 @@ export default {
       if (context.url.startsWith('/src/index.map.js')) {
         await next();
         context.body = context.body
-          .replace(/navigator\.sendBeacon/g, 'fakeSendBeacon');
+          .replace(/navigator\.sendBeacon/g, 'fakeSendBeacon')
+          // rewriting dynamic plugins base url
+          .replace(/document\.currentScript\.src/g, '"http://localhost:8000/plugins"');
+        return true;
+      } else if (context.url.startsWith('/src/index.js')
+          || context.url.startsWith('/modules/index.js')) {
+        await next();
+        context.body = context.body
+          // rewriting dynamic plugins base url
+          .replace(/document\.currentScript\.src/g, '"http://localhost:8000/plugins"');
+        return true;
+      } else if (context.url.startsWith('/modules/index-broken.js')) {
+        const [_, search] = context.url.split('?');
+        context.url = `/modules/index.js?${search}`;
+        await next();
+        context.body = context.body
+          // rewriting dynamic plugins base url
+          .replace(/document\.currentScript\.src/g, '"http://localhost:8000/plugins"')
+          .replace(/\/\/ test: broken-plugin/g, 'foo: "foo.js",');
         return true;
       } else if (context.url.startsWith('/.rum')) {
         if (context.url.startsWith('/.rum/@adobe/helix-rum-js@%5E2/dist/')
@@ -43,9 +74,10 @@ export default {
           context.url = '/node_modules/@adobe/helix-rum-js/dist/rum-standalone.js';
           await next();
           context.body = context.body
-            .replace(/const weight.*/, 'const weight = 1;')
+            .replace(/const weight =/, 'const weight = 1 ||')
             .replace(/navigator\.sendBeacon/g, 'fakeSendBeacon')
-            .replace('.rum/@adobe/helix-rum-enhancer@^2/src/index.js', 'src/index.map.js');
+            // eslint-disable-next-line no-template-curly-in-string
+            .replace('.rum/@adobe/helix-rum-enhancer@${enhancerVersion || \'^2\'}/src/index.js', 'src/index.map.js');
           return true;
         } else if (context.url.startsWith('/.rum/web-vitals')) {
           context.url = '/node_modules/web-vitals/dist/web-vitals.iife.js';